+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Camera/PlayerCameraManager.h
@@ -15,16 +15,16 @@ class APlayerCameraManager : public AActor
 public:
     APlayerCameraManager();
 
-    FLinearColor FadeColor;
-    float FadeAmount;
-    FVector2D FadeAlpha;
-    float FadeTime;
-    float FadeTimeRemaining;
+    FLinearColor FadeColor; // 목표 fade 색상
+    float FadeAmount; // 현재 fade 얼마나되있는지
+    FVector2D FadeAlpha; // 현재 실행되는 fade 효과의 시작 끝 정도
+    float FadeTime; // 현재 실행되는 fade 효과의 시간
+    float FadeTimeRemaining; // 남은 fade 효과의 시간
 
     FName CameraStyle; // ?
     //struct FViewTarget ViewTarget; -> 뷰포트 직접 참조로 대체
 
-    void UpdateCamera(float DeltaTime);
+    void DoUpdateCamera(float DeltaTime);
     UCameraModifier* AddCameraModifier(UCameraModifier* InModifier);
     void RemoveCameraModifier(UCameraModifier* InModifier);
     void ClearCameraModifiers();
@@ -33,6 +33,7 @@ public:
     void DisableModifier(UCameraModifier* InModifier);
 
     void StartCameraFade(float FromAlpha, float ToAlpha, float Duration, const FLinearColor& Color);
+    void StopCameraFade();
     UCameraComponent* GetCameraComponent();
     void SetCameraComponent(UCameraComponent* InCameraComponent);
