+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/CompositingPass.h
@@ -1,4 +1,4 @@
-﻿
+
 #pragma once
 #include <memory>
