--- EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/LineRenderPass.cpp
@@ -1,129 +1,0 @@
-// DEPRECATED
-#if 0
-#include "LineRenderPass.h"
-
-#include "D3D11RHI/DXDBufferManager.h"
-
-#include "D3D11RHI/GraphicDevice.h"
-
-#include "D3D11RHI/DXDShaderManager.h"
-
-#include "RendererHelpers.h"
-
-#include "Math/JungleMath.h"
-
-#include "EngineLoop.h"
-#include "UnrealClient.h"
-
-#include "UObject/UObjectIterator.h"
-
-#include "UnrealEd/EditorViewportClient.h"
-
-// 생성자/소멸자
-FLineRenderPass::FLineRenderPass()
-    : BufferManager(nullptr)
-    , Graphics(nullptr)
-    , ShaderManager(nullptr)
-    , VertexLineShader(nullptr)
-    , PixelLineShader(nullptr)
-{
-}
-
-FLineRenderPass::~FLineRenderPass()
-{
-}
-
-void FLineRenderPass::Initialize(FDXDBufferManager* InBufferManager, FGraphicsDevice* InGraphics, FDXDShaderManager* InShaderManager)
-{
-    BufferManager = InBufferManager;
-    Graphics = InGraphics;
-    ShaderManager = InShaderManager;
-    CreateShader();
-}
-
-void FLineRenderPass::PrepareRender()
-{
-}
-
-void FLineRenderPass::ClearRenderArr()
-{
-    // 필요에 따라 내부 배열을 초기화
-}
-
-void FLineRenderPass::CreateShader()
-{
-    HRESULT hr = ShaderManager->AddVertexShader(L"VertexLineShader", L"Shaders/ShaderLine.hlsl", "mainVS");
-    hr = ShaderManager->AddPixelShader(L"PixelLineShader", L"Shaders/ShaderLine.hlsl", "mainPS");
-
-    VertexLineShader = ShaderManager->GetVertexShaderByKey(L"VertexLineShader");
-    PixelLineShader = ShaderManager->GetPixelShaderByKey(L"PixelLineShader");
-}
-
-void FLineRenderPass::PrepareLineShader() const
-{
-    Graphics->DeviceContext->VSSetShader(VertexLineShader, nullptr, 0);
-    Graphics->DeviceContext->PSSetShader(PixelLineShader, nullptr, 0);
-
-    FEngineLoop::PrimitiveDrawBatch.PrepareLineResources();
-}
-
-void FLineRenderPass::DrawLineBatch(const FLinePrimitiveBatchArgs& BatchArgs) const
-{
-    UINT stride = sizeof(FSimpleVertex);
-    UINT offset = 0;
-    Graphics->DeviceContext->IASetVertexBuffers(0, 1, &BatchArgs.VertexBuffer, &stride, &offset);
-    Graphics->DeviceContext->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_LINELIST);
-
-    const UINT vertexCountPerInstance = 2;
-    UINT instanceCount = BatchArgs.GridParam.NumGridLines + 3 +
-        (BatchArgs.BoundingBoxCount * 12) +
-        (BatchArgs.ConeCount * (2 * BatchArgs.ConeSegmentCount)) +
-        (12 * BatchArgs.OBBCount);
-
-    Graphics->DeviceContext->DrawInstanced(vertexCountPerInstance, instanceCount, 0, 0);
-    Graphics->DeviceContext->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
-}
-
-void FLineRenderPass::ReloadShader()
-{
-    VertexLineShader = ShaderManager->GetVertexShaderByKey(L"VertexLineShader");
-    PixelLineShader = ShaderManager->GetPixelShaderByKey(L"PixelLineShader");
-}
-
-void FLineRenderPass::UpdateObjectConstant(const FMatrix& WorldMatrix, const FVector4& UUIDColor, bool bIsSelected) const
-{
-    FObjectConstantBuffer ObjectData = {};
-    ObjectData.WorldMatrix = WorldMatrix;
-    ObjectData.InverseTransposedWorld = FMatrix::Transpose(FMatrix::Inverse(WorldMatrix));
-    ObjectData.UUIDColor = UUIDColor;
-    ObjectData.bIsSelected = bIsSelected;
-    
-    BufferManager->UpdateConstantBuffer(TEXT("FObjectConstantBuffer"), ObjectData);
-}
-
-void FLineRenderPass::ProcessLineRendering(const std::shared_ptr<FEditorViewportClient>& Viewport)
-{
-    PrepareLineShader();
-
-    // 상수 버퍼 업데이트: Identity 모델, 기본 색상 등
-    UpdateObjectConstant(FMatrix::Identity, FVector4(0, 0, 0, 0), false);
-
-    FLinePrimitiveBatchArgs BatchArgs;
-    FEngineLoop::PrimitiveDrawBatch.PrepareBatch(BatchArgs);
-    DrawLineBatch(BatchArgs);
-    FEngineLoop::PrimitiveDrawBatch.RemoveArr();
-}
-
-void FLineRenderPass::Render(const std::shared_ptr<FEditorViewportClient>& Viewport)
-{
-    const EResourceType ResourceType = EResourceType::ERT_Editor;
-
-    FViewportResource* ViewportResource = Viewport->GetViewportResource();
-    FRenderTargetRHI* RenderTargetRHI = ViewportResource->GetRenderTarget(ResourceType);
-    Graphics->DeviceContext->OMSetRenderTargets(1, &RenderTargetRHI->RTV, ViewportResource->GetDepthStencilView());
-
-    ProcessLineRendering(Viewport);
-
-    Graphics->DeviceContext->OMSetRenderTargets(0, nullptr, nullptr);
-}
-#endif
