+++ EngineSIU/EngineSIU/editor.ini
@@ -1,5 +1,5 @@
 SplitterH.SplitRatio=0.500000
-PerspectiveCameraLocY0=67.048790
+PerspectiveCameraLocY0=-8.049848
 PerspectiveCameraLocX3=0.000000
 PerspectiveCameraRotY1=0.000000
 SplitterV.SplitRatio=0.500000
@@ -8,15 +8,15 @@ ShowFlag1=31
 CameraSpeedSetting0=1
 CameraSpeedScalar0=1.378222
 PerspectiveCameraRotY2=0.000000
-PerspectiveCameraLocX0=-82.661316
+PerspectiveCameraLocX0=-23.237524
 GridSize0=10.000000
-PerspectiveCameraLocZ0=85.620193
+PerspectiveCameraLocZ0=15.231767
 ShowFlag3=31
 CameraSpeedSetting2=1
 PerspectiveCameraRotX0=0.000000
 PerspectiveCameraLocX2=0.000000
-PerspectiveCameraRotY0=37.799774
-PerspectiveCameraRotZ0=-1816.938721
+PerspectiveCameraRotY0=37.699768
+PerspectiveCameraRotZ0=-1784.838257
 PerspectiveCameraRotX3=0.000000
 CameraSpeedSetting1=1
 ShowFlag0=30
