+++ EngineSIU/EngineSIU/imgui.ini
@@ -11,8 +11,8 @@ Pos=1032,303
 Size=250,624
 
 [Window][Console]
-Pos=0,0
-Size=32,34
+Pos=0,576
+Size=642,384
 
 [Window][Control Panel]
 Pos=1,1
